ver2

+++ Terz/Controllers/DataManagerController.cs
@@ -3,8 +3,10 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Newtonsoft.Json;
+using Terz_DataBaseLayer;
 
 namespace Terz.Controllers
 {
@@ -12,6 +14,16 @@ namespace Terz.Controllers
     {
         public PartialViewResult Index([FromQuery(Name = "id")] string id)
         {
+            string userId = HttpContext.Session.GetString("User");
+
+            if (userId == null || userId == "")
+            {
+                return PartialView("~/Views/Home/Login.cshtml");
+            }
+            if (!Security.CheckReportPermission(userId, id))
+            {
+                return PartialView("~/Views/Home/Login.cshtml");
+            }
             Models.DataManager.DataManagerView dataManagerView = new Models.DataManager.DataManagerView();
             dataManagerView.Id = id;
             List<Models.DataManager.DataFrameInfo> dataFrameInfos = new List<Models.DataManager.DataFrameInfo>();
@@ -29,11 +41,40 @@ namespace Terz.Controllers
                 dataFrameInfos.Add(dataFrameInfo);
             }
             dataManagerView.DataFrameInfos = dataFrameInfos;
+            dataManagerView.Report = new Report();
+            dataManagerView.Report.Load(id);
+            dataManagerView.Report.getReference();
+            
             return PartialView(dataManagerView);
+        }
+        public string AddReferencia(Terz_DataBaseLayer.Referencia referencia)
+        {
+            string userId = HttpContext.Session.GetString("User");
+
+            if (userId == null || userId == "")
+            {
+                return "not authenticated";
+            }
+            if (!Security.CheckReportPermission(userId, referencia.ReportId))
+            {
+                return "no permission";
+            }
+            referencia.Insert();
+            return "ok";
         }
 
         public async Task<string> UploadDataFrame([FromQuery(Name = "id")] string id)
         {
+            string userId = HttpContext.Session.GetString("User");
+
+            if (userId == null || userId == "")
+            {
+                return "not authenticated";
+            }
+            if (!Security.CheckReportPermission(userId, id))
+            {
+                return "no permission";
+            }
             string text = System.IO.File.ReadAllText(Location.ConfLocation);
             Conf conf = JsonConvert.DeserializeObject<Conf>(text);
 
@@ -51,6 +92,16 @@ namespace Terz.Controllers
 
         public string RenameDataFrame([FromQuery(Name = "id")] string id, [FromQuery(Name = "name")] string name, [FromQuery(Name = "newname")] string newName)
         {
+            string userId = HttpContext.Session.GetString("User");
+
+            if (userId == null || userId == "")
+            {
+                return "not authenticated";
+            }
+            if (!Security.CheckReportPermission(userId, id))
+            {
+                return "no permission";
+            }
             string text = System.IO.File.ReadAllText(Location.ConfLocation);
             Conf conf = JsonConvert.DeserializeObject<Conf>(text);
             var oldFile = Path.Combine(conf.DataFramePath + "/" + id, name+".csv");
@@ -59,12 +110,60 @@ namespace Terz.Controllers
             return "ok";
         }
 
+        public string RenameReferencia([FromQuery(Name = "id")] string id, [FromQuery(Name = "desc")] string desc)
+        {
+            string userId = HttpContext.Session.GetString("User");
+
+            if (userId == null || userId == "")
+            {
+                return "not authenticated";
+            }
+            if (!Security.CheckReferencePermission(userId, id))
+            {
+                return "no permission";
+            }
+            Terz_DataBaseLayer.Referencia referencia = new Referencia();
+            referencia.Id = id;
+            referencia.Descricao = desc;
+            referencia.Update();
+
+            return "ok";
+        }
+
         public string DeleteDataFrame([FromQuery(Name = "id")] string id, [FromQuery(Name = "name")] string name)
         {
+            string userId = HttpContext.Session.GetString("User");
+
+            if (userId == null || userId == "")
+            {
+                return "not authenticated";
+            }
+            if (!Security.CheckReportPermission(userId, id))
+            {
+                return "no permission";
+            }
             string text = System.IO.File.ReadAllText(Location.ConfLocation);
             Conf conf = JsonConvert.DeserializeObject<Conf>(text);
             var file = Path.Combine(conf.DataFramePath + "/" + id, name + ".csv");
             System.IO.File.Delete(file);
+            return "ok";
+        }
+
+        public string DeleteReferencia([FromQuery(Name = "id")] string id)
+        {
+            string userId = HttpContext.Session.GetString("User");
+
+            if (userId == null || userId == "")
+            {
+                return "not authenticated";
+            }
+            if (!Security.CheckReferencePermission(userId, id))
+            {
+                return "no permission";
+            }
+            Terz_DataBaseLayer.Referencia referencia = new Referencia();
+            referencia.Id = id;
+            referencia.Delete();
             return "ok";
         }
     }
